[feature/extend-src] merge dev
