+++ src/modules/home/components/Content.vue
@@ -1,61 +1,63 @@
 <template>
-  <CardWrapper btn-body-class="col-lg-24">
-    <div class="content">
-      
-      <div class="row justify-content-center p-3">
-        <div class="d-flex container bg-white wapper">
-          <div class="flex-grow-1" style="margin: unset">
-            <h5 class="text-left">doLogin - onclick - id::btn_login</h5>
-            <!-- Title for the table -->
-            <table class="table table-bordered mt-3">
-              <thead class="table-dark">
-                <tr>
-                  <th>#</th>
-                  <th>Category</th>
-                  <th>Type</th>
-                  <th>CElement</th>
-                  <th>Id / Name / Class</th>
-                </tr>
-              </thead>
-              <tbody>
-                <tr>
-                  <td>1</td>
-                  <td>input</td>
-                  <td>text</td>
-                  <td>id</td>
-                  <td>user_id</td>
-                </tr>
-                <tr>
-                  <td>1</td>
-                  <td>input</td>
-                  <td>password</td>
-                  <td>id</td>
-                  <td>password</td>
-                </tr>
-              </tbody>
-            </table>
-            <div class="text-left mt-3">
-              <button class="btn btn-primary" @click="toggleModal">SELECT VIEW POINT</button>
-              <Modal :showModal="showModal" :toggleModal="toggleModal" />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- Block Footer-->
-    <Footer></Footer>
+  <CardWrapper
+      v-for="([key, value], index) in mapEvent.entries()"
+      :key="index" :class="index === mapEvent.size - 1 ? 'mb-3' : ''"
+  >
+    <TableEvent
+        :title="key"
+        :headers="headersForTable"
+        :data="value"
+        :field-table="fieldTable"
+    />
   </CardWrapper>
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
-import Footer from "./Footer.vue";
-import Modal from "./Modal.vue";
-const showModal = ref(false);
-const toggleModal = () => {
-    showModal.value = !showModal.value;
-};
+import CardWrapper from "@/components/common/card/CardWrapper.vue";
+import TableEvent from "@/modules/home/components/table/TableEvent.vue";
+import {useTable} from "@/components/utils/table-utils";
+import type {ITableEvent} from "@/modules/home/home.type";
+import type {IHeaderTable} from "@/components/common/table/header/TableHeader.type";
+
+defineProps<{
+  mapEvent: Map<string, ITableEvent[]>
+}>()
+const headersForTable: IHeaderTable[][] = [
+  [
+    {
+      text: '#',
+      value: 'serial',
+      align: 'center',
+      width: 60
+    },
+    {
+      text: 'Category',
+      value: 'category',
+      align: 'start',
+      width: 200
+    },
+    {
+      text: 'Type',
+      value: 'type',
+      align: 'start',
+      width: 200
+    },
+    {
+      text: 'CElement',
+      value: 'c_element',
+      align: 'start',
+      width: 200
+    },
+    {
+      text: 'Id / Name / Class',
+      value: 'selector',
+      align: 'start',
+      width: 200
+    }
+  ]
+]
+
+const fieldTable = useTable(headersForTable)
 
 </script>
